refactor: move e2e tests to cli dir
